Now OBO netman runs with threads
There was also a test added to ensure gunicorn wrapping compatibility
