+++ test-requirements.txt
@@ -5,4 +5,6 @@ pyhamcrest>=1.6
 MockSSH>=1.4.1
 Sphinx
 sphinxcontrib-httpdomain
-fake-switches>=1.0.14
+fake-switches>=1.0.15
+gunicorn>=19.4.5
+futures>=3.0.4
